+++ web_module/trunk/src/main/com/topcoder/web/common/dao/UserPreferenceDAO.java
@@ -11,8 +11,10 @@ import com.topcoder.web.common.model.UserPreference;
  */
 public interface UserPreferenceDAO {
 
-	UserPreference find(User user, Preference preference);
-	UserPreference find(Long userId, Long preferenceId);	
-	void saveOrUpdate(UserPreference u);
+    UserPreference find(User user, Preference preference);
+
+    UserPreference find(Long userId, Integer preferenceId);
+
+    void saveOrUpdate(UserPreference u);
 
 }
